requirements.txt
tqdm
scikit-learn
torchvision
torch
attrs==21.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
ci-info==0.2.0
click==8.0.1
configobj==5.0.6
configparser==5.0.2
dataclasses==0.6
decorator==4.4.2
etelemetry==0.2.2
filelock==3.0.12
fitz==0.0.1.dev2
Flask==2.0.1
future==0.18.2
httplib2==0.19.1
idna==2.10
importlib-metadata==4.3.0
inflection==0.5.1
isodate==0.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
marshmallow==3.12.1
networkx==2.5.1
nibabel==3.2.1
nipype==1.6.0
numpy==1.20.3
packaging==20.9
pandas==1.2.4
patent-client==2.2.1
pathlib==1.0.1
Pillow==8.2.0
prov==2.0.0
pydot==1.4.2
PyMuPDF==1.18.13
pyparsing==2.4.7
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2021.1
pyxnat==1.4
rdflib==5.0.0
redis==3.5.3
requests==2.25.1
requests-cache==0.6.3
scipy==1.6.3
simplejson==3.17.2
six==1.16.0
soupsieve==2.2.1
traits==6.2.0
typing-extensions==3.10.0.0
url-normalize==1.4.3
urllib3==1.26.5
Werkzeug==2.0.1
xmltodict==0.12.0
zipp==3.4.1
python-dotenv==0.17.1
celery==5.0.5
